pyproject.toml
[project]
name = "ait-app"
version = "1.0.0"
description = "AiT (Anonymization in Time) - Video anonymization tool for chat recordings"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "numpy==1.26.4",
    "Pillow>=10.0.0",
    "opencv-python>=4.8.0",
    "tqdm>=4.66.0",
    "joblib>=1.3.0",
]

[project.optional-dependencies]
processing = [
    "easyocr>=1.7.1",
    "einops>=0.6.1",
    # Torch pins (use platform-appropriate wheels; see README for CUDA/MPS notes)
    "torch==2.7.0",
    "torchvision==0.22.0",
    "torchaudio==2.7.0",
    "transformers>=4.46.0",
]
build = [
    "pyinstaller>=6.0.0",
]

[project.scripts]
ait = "ait.launcher:main"
ait-process = "ait.process_videos:main"
ait-viewer = "ait.annotation_viewer:main"
ait-export = "ait.export_video:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ait"]
